Moved validation to testcase
