--- src/main/java/app/App.java
@@ -1,5 +1,0 @@
-package app;
-
-public class App {
-
-}
