--- src/test/java/app/TestApp.java
@@ -1,5 +1,0 @@
-package app;
-
-public class TestApp {
-
-}
